+++ pkg/authz/check/service.go
@@ -48,6 +48,7 @@ func NewService(env service.Env, warrantRepo warrant.WarrantRepository, eventSvc
 func (svc CheckService) getWithPolicyMatch(ctx context.Context, sema *Semaphore, spec CheckWarrantSpec) (*warrant.WarrantSpec, error) {
 	sema.Acquire()
 	defer sema.Release()
+
 	warrants, err := svc.WarrantRepository.GetAllMatchingObjectRelationAndSubject(ctx, spec.ObjectType, spec.ObjectId, spec.Relation, spec.Subject.ObjectType, spec.Subject.ObjectId, spec.Subject.Relation)
 	if err != nil || len(warrants) == 0 {
 		return nil, err
@@ -74,7 +75,6 @@ func (svc CheckService) getWithPolicyMatch(ctx context.Context, sema *Semaphore,
 func (svc CheckService) getMatchingSubjects(ctx context.Context, sema *Semaphore, objectTypeMap *objecttype.ObjectTypeMap, objectType string, objectId string, relation string, checkCtx warrant.PolicyContext) ([]warrant.WarrantSpec, error) {
 	sema.Acquire()
 	defer sema.Release()
-	//log.Ctx(ctx).Debug().Msgf("Getting matching subjects for %s:%s#%s@___%s", objectType, objectId, relation, checkCtx)
 
 	warrantSpecs := make([]warrant.WarrantSpec, 0)
 	objectTypeSpec, err := objectTypeMap.GetByTypeId(objectType)
@@ -114,8 +114,6 @@ func (svc CheckService) getMatchingSubjects(ctx context.Context, sema *Semaphore
 }
 
 func (svc CheckService) getMatchingSubjectsBySubjectType(ctx context.Context, sema *Semaphore, objectTypeMap *objecttype.ObjectTypeMap, objectType string, objectId string, relation string, subjectType string, checkCtx warrant.PolicyContext) ([]warrant.WarrantSpec, error) {
-	// log.Ctx(ctx).Debug().Msgf("Getting matching subjects for %s:%s#%s@%s:___%s", objectType, objectId, relation, subjectType, checkCtx)
-
 	sema.Acquire()
 	defer sema.Release()
 
@@ -331,11 +329,11 @@ func (svc CheckService) Check(ctx context.Context, authInfo *service.AuthInfo, w
 		return false, nil, nil, err
 	}
 	resultsC := make(chan Result)
-	sema := NewSema(8)
+	sema := NewSema(4)
 	// TODO: should use a cancellable ctx here (with timeout)
 	// TODO: Should do wookieSafeRead
 	go func() {
-		check(0, sema, ctx, warrantCheck, make([]warrant.WarrantSpec, 0), resultsC, &svc, typeMap)
+		svc.check(0, sema, ctx, warrantCheck, make([]warrant.WarrantSpec, 0), resultsC, typeMap)
 	}()
 	result := <-resultsC
 	if result.Err != nil {
@@ -345,6 +343,217 @@ func (svc CheckService) Check(ctx context.Context, authInfo *service.AuthInfo, w
 		return true, result.DecisionPath, nil, nil
 	}
 	return false, nil, nil, nil
+}
+
+// TODO: fix decisionPath
+func (svc CheckService) check(level int, sema *Semaphore, ctx context.Context, checkSpec CheckSpec, currentPath []warrant.WarrantSpec, resultC chan<- Result, typesMap *objecttype.ObjectTypeMap) {
+	select {
+	case <-ctx.Done():
+		log.Ctx(ctx).Debug().Msgf("canceled check[%d] [%s]", level, checkSpec)
+		return
+	default:
+		log.Ctx(ctx).Debug().Msgf("exec check[%d] [%s]", level, checkSpec)
+		// 1. Check for direct warrant match
+		matchedWarrant, err := svc.getWithPolicyMatch(ctx, sema, checkSpec.CheckWarrantSpec)
+		if err != nil {
+			// log.Ctx(ctx).Err(err).Msgf("ERR CheckTask getWithPolicyMatch")
+			resultC <- Result{
+				Matched:      false,
+				DecisionPath: currentPath,
+				Err:          err,
+			}
+			return
+		}
+		if matchedWarrant != nil {
+			resultC <- Result{
+				Matched:      true,
+				DecisionPath: append(currentPath, *matchedWarrant),
+				Err:          nil,
+			}
+			return
+		}
+
+		// 2. Check through indirect/group warrants
+		var additionalTasks []func(execCtx context.Context, resultC chan<- Result)
+		additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+			svc.checkGroup(level+1, sema, execCtx, checkSpec, currentPath, resultC, typesMap)
+		})
+
+		// 3. And/or defined rules for target relation
+		objectTypeSpec, err := typesMap.GetByTypeId(checkSpec.ObjectType)
+		if err != nil {
+			//log.Ctx(ctx).Err(err).Msgf("ERR CheckTask GetByTypeId")
+			resultC <- Result{
+				Matched:      false,
+				DecisionPath: currentPath,
+				Err:          err,
+			}
+			return
+		}
+		if relationRule, ok := objectTypeSpec.Relations[checkSpec.Relation]; ok {
+			additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+				svc.checkRule(level+1, sema, execCtx, checkSpec, currentPath, resultC, typesMap, &relationRule)
+			})
+		}
+
+		resultC <- anyOfBlocking(ctx, additionalTasks)
+	}
+}
+
+func (svc CheckService) checkGroup(level int, sema *Semaphore, ctx context.Context, checkSpec CheckSpec, currentPath []warrant.WarrantSpec, resultC chan<- Result, typesMap *objecttype.ObjectTypeMap) {
+	select {
+	case <-ctx.Done():
+		log.Ctx(ctx).Debug().Msgf("canceled checkGroup[%d] [%s]", level, checkSpec)
+		return
+	default:
+		log.Ctx(ctx).Debug().Msgf("exec checkGroup[%d] [%s]", level, checkSpec)
+		warrants, err := svc.getMatchingSubjects(ctx, sema, typesMap, checkSpec.ObjectType, checkSpec.ObjectId, checkSpec.Relation, checkSpec.Context)
+		if err != nil {
+			//log.Ctx(ctx).Err(err).Msgf("ERR GroupCheckTask getMatchingSubjects")
+			resultC <- Result{
+				Matched:      false,
+				DecisionPath: currentPath,
+				Err:          err,
+			}
+			return
+		}
+
+		var matchingWarrants []warrant.WarrantSpec
+		for _, w := range warrants {
+			if w.Subject.Relation == "" {
+				continue
+			}
+			matchingWarrants = append(matchingWarrants, w)
+		}
+		if len(matchingWarrants) == 0 {
+			resultC <- Result{
+				Matched:      false,
+				DecisionPath: currentPath,
+				Err:          nil,
+			}
+		}
+		var additionalTasks []func(execCtx context.Context, resultC chan<- Result)
+		for _, w := range matchingWarrants {
+			matchingWarrant := w
+			additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+				svc.check(level+1, sema, execCtx, CheckSpec{
+					CheckWarrantSpec: CheckWarrantSpec{
+						ObjectType: matchingWarrant.Subject.ObjectType,
+						ObjectId:   matchingWarrant.Subject.ObjectId,
+						Relation:   matchingWarrant.Subject.Relation,
+						Subject:    checkSpec.Subject,
+						Context:    checkSpec.Context,
+					},
+					Debug: checkSpec.Debug,
+				}, append(currentPath, matchingWarrant), resultC, typesMap)
+			})
+		}
+		resultC <- anyOfBlocking(ctx, additionalTasks)
+	}
+}
+
+func (svc CheckService) checkRule(level int, sema *Semaphore, ctx context.Context, checkSpec CheckSpec, currentPath []warrant.WarrantSpec, resultC chan<- Result, typesMap *objecttype.ObjectTypeMap, rule *objecttype.RelationRule) {
+	select {
+	case <-ctx.Done():
+		log.Ctx(ctx).Debug().Msgf("canceled checkRule[%d] [%s] [%s]", level, checkSpec, rule)
+		return
+	default:
+		log.Ctx(ctx).Debug().Msgf("exec checkRule[%d] [%s] [%s]", level, checkSpec, rule)
+		warrantSpec := checkSpec.CheckWarrantSpec
+		if rule == nil {
+			resultC <- Result{
+				Matched:      false,
+				DecisionPath: currentPath,
+				Err:          nil,
+			}
+			return
+		}
+		switch rule.InheritIf {
+		case "":
+			// No match found
+			resultC <- Result{
+				Matched:      false,
+				DecisionPath: currentPath,
+				Err:          nil,
+			}
+		case objecttype.InheritIfAllOf:
+			var additionalTasks []func(execCtx context.Context, resultC chan<- Result)
+			for _, r := range rule.Rules {
+				subRule := r
+				additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+					svc.checkRule(level+1, sema, execCtx, checkSpec, currentPath, resultC, typesMap, &subRule)
+				})
+			}
+			resultC <- allOfBlocking(ctx, additionalTasks)
+		case objecttype.InheritIfAnyOf:
+			var additionalTasks []func(execCtx context.Context, resultC chan<- Result)
+			for _, r := range rule.Rules {
+				subRule := r
+				additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+					svc.checkRule(level+1, sema, execCtx, checkSpec, currentPath, resultC, typesMap, &subRule)
+				})
+			}
+			resultC <- anyOfBlocking(ctx, additionalTasks)
+		case objecttype.InheritIfNoneOf:
+			var additionalTasks []func(execCtx context.Context, resultC chan<- Result)
+			for _, r := range rule.Rules {
+				subRule := r
+				additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+					svc.checkRule(level+1, sema, execCtx, checkSpec, currentPath, resultC, typesMap, &subRule)
+				})
+			}
+			resultC <- noneOfBlocking(ctx, additionalTasks)
+		default:
+			if rule.OfType == "" && rule.WithRelation == "" {
+				svc.check(level+1, sema, ctx, CheckSpec{
+					CheckWarrantSpec: CheckWarrantSpec{
+						ObjectType: warrantSpec.ObjectType,
+						ObjectId:   warrantSpec.ObjectId,
+						Relation:   rule.InheritIf,
+						Subject:    warrantSpec.Subject,
+						Context:    warrantSpec.Context,
+					},
+					Debug: checkSpec.Debug,
+				}, currentPath, resultC, typesMap)
+				return
+			}
+
+			matchingWarrants, err := svc.getMatchingSubjectsBySubjectType(ctx, sema, typesMap, warrantSpec.ObjectType, warrantSpec.ObjectId, rule.WithRelation, rule.OfType, warrantSpec.Context)
+			if err != nil {
+				//log.Ctx(ctx).Err(err).Msgf("ERR CheckRuleTask getMatchingSubjectsBySubjectType")
+				resultC <- Result{
+					Matched:      false,
+					DecisionPath: currentPath,
+					Err:          err,
+				}
+				return
+			}
+			if len(matchingWarrants) == 0 {
+				resultC <- Result{
+					Matched:      false,
+					DecisionPath: currentPath,
+					Err:          nil,
+				}
+			}
+			var additionalTasks []func(execCtx context.Context, resultC chan<- Result)
+			for _, w := range matchingWarrants {
+				matchingWarrant := w
+				additionalTasks = append(additionalTasks, func(execCtx context.Context, resultC chan<- Result) {
+					svc.check(level+1, sema, execCtx, CheckSpec{
+						CheckWarrantSpec: CheckWarrantSpec{
+							ObjectType: matchingWarrant.Subject.ObjectType,
+							ObjectId:   matchingWarrant.Subject.ObjectId,
+							Relation:   rule.InheritIf,
+							Subject:    warrantSpec.Subject,
+							Context:    warrantSpec.Context,
+						},
+						Debug: checkSpec.Debug,
+					}, append(currentPath, matchingWarrant), resultC, typesMap)
+				})
+			}
+			resultC <- anyOfBlocking(ctx, additionalTasks)
+		}
+	}
 }
 
 func evalWarrantPolicy(w warrant.Model, policyCtx warrant.PolicyContext) bool {
